fix: float total measured

+++ src/types/serviceOrder.ts
@@ -40,7 +40,7 @@ export const ServiceOrderSchema = z.object({
   minimal_documentation: z.array(z.string()).optional(),
   pls_verifications: z.array(z.string()).optional(),
   pls_built_situation: z.string().optional(),
-  total_measured: z.number().int().optional(),
+  total_measured: z.number().optional(),
   more_accurate_informations: z.array(z.string()).optional(),
   mandatory_documents_to_b: z.array(z.string()).optional(),
   pci_verifications: z.array(z.string()).optional(),
